requirements.txt
opencv_python
tqdm
numpy
albumentations
torch
scipy
torchvision
efficientnet_pytorch
inplace_abn
nni
Pillow
pretrainedmodels
timm
yacs
